src/app/requirements.txt
fastapi[standard]==0.115.0
SQLAlchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
aiohttp==3.10.5
python-dotenv==1.0.1
# pydantic-settings==2.5.2
